调整后台管理系统服务，与前端数据一致

+++ ims-service/src/main/java/com/lhiot/ims/rbac/mapper/ImsUserMapper.java
@@ -80,5 +80,5 @@ public interface ImsUserMapper {
     * @author yijun
     * @date 2018/09/29 11:42:57
     */
-    long pageImsUserCounts(ImsUser imsUser);
+    int pageImsUserCounts(ImsUser imsUser);
 }
